Fixed a lot of non-standard conforming code.

+++ PoleBalancingSFML/Source/deep/DSOM.cpp
@@ -192,7 +192,7 @@ void DSOM::updateNeighborhood(const SOMCoords &centerCoords, const std::vector<f
 
 		float dist = std::sqrt(distSquared);
 
-		float influence = (_neighborhoodRadius - dist) / _neighborhoodRadius * std::expf(-_gaussianScalar * distSquared / (2.0f * radiusSquaredf));
+		float influence = (_neighborhoodRadius - dist) / _neighborhoodRadius * std::exp(-_gaussianScalar * distSquared / (2.0f * radiusSquaredf));
 
 		for (size_t i = 0; i < node._weights.size(); i++)
 			node._weights[i] += influence * _alpha * (target[i] - node._weights[i]);
