+++ PoleBalancingSFML/Source/lstm/LSTMG.h
@@ -97,7 +97,7 @@ namespace lstm {
 		};
 
 		static float sigmoid(float x) {
-			return 1.0f / (1.0f + std::expf(-x));
+			return 1.0f / (1.0f + std::exp(-x));
 		}
 
 		static float sigmoidDerivative(float x) {
