+++ PoleBalancingSFML/Source/lstm/LSTMNet.h
@@ -69,7 +69,7 @@ namespace lstm {
 		};
 
 		static float sigmoid(float x) {
-			return 1.0f / (1.0f + std::expf(-x));
+			return 1.0f / (1.0f + std::exp(-x));
 		}
 
 	private:
