+++ PoleBalancingSFML/Source/lstmrl/LSTMRL.h
@@ -85,7 +85,7 @@ namespace lstmrl {
 		};
 
 		static float sigmoid(float x) {
-			return 1.0f / (1.0f + std::expf(-x));
+			return 1.0f / (1.0f + std::exp(-x));
 		}
 
 	private:
